Move existing usage of task data to envrionment impl
