+++ SpiffWorkflow/bpmn/PythonScriptEngine.py
@@ -4,6 +4,7 @@ import copy
 import sys
 import traceback
 
+from .PythonScriptEngineEnvironment import TaskDataEnvironment
 from ..exceptions import SpiffWorkflowException, WorkflowTaskException
 from ..operators import Operator
 
@@ -26,66 +27,6 @@ from ..operators import Operator
 # 02110-1301  USA
 
 
-class Box(dict):
-    """
-    Example:
-    m = Box({'first_name': 'Eduardo'}, last_name='Pool', age=24, sports=['Soccer'])
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(Box, self).__init__(*args, **kwargs)
-        for arg in args:
-            if isinstance(arg, dict):
-                for k, v in arg.items():
-                    if isinstance(v, dict):
-                        self[k] = Box(v)
-                    else:
-                        self[k] = v
-
-        if kwargs:
-            for k, v in kwargs.items():
-                if isinstance(v, dict):
-                    self[k] = Box(v)
-                else:
-                    self[k] = v
-
-    def __deepcopy__(self, memodict=None):
-        if memodict is None:
-            memodict = {}
-        my_copy = Box()
-        for k, v in self.items():
-            my_copy[k] = copy.deepcopy(v)
-        return my_copy
-
-    def __getattr__(self, attr):
-        try:
-            output = self[attr]
-        except:
-            raise AttributeError(
-                "Dictionary has no attribute '%s' " % str(attr))
-        return output
-
-    def __setattr__(self, key, value):
-        self.__setitem__(key, value)
-
-    def __setitem__(self, key, value):
-        super(Box, self).__setitem__(key, value)
-        self.__dict__.update({key: value})
-
-    def __getstate__(self):
-        return self.__dict__
-
-    def __setstate__(self, state):
-        self.__init__(state)
-
-    def __delattr__(self, item):
-        self.__delitem__(item)
-
-    def __delitem__(self, key):
-        super(Box, self).__delitem__(key)
-        del self.__dict__[key]
-
-
 class PythonScriptEngine(object):
     """
     This should serve as a base for all scripting & expression evaluation
@@ -97,10 +38,14 @@ class PythonScriptEngine(object):
     expressions in a different way.
     """
 
-    def __init__(self, default_globals=None, scripting_additions=None):
-
-        self.globals = default_globals or {}
-        self.globals.update(scripting_additions or {})
+    def __init__(self, default_globals=None, scripting_additions=None, environment=None):
+        if environment is None:
+            environment_globals = {}
+            environment_globals.update(default_globals or {})
+            environment_globals.update(scripting_additions or {})
+            self.environment = TaskDataEnvironment(environment_globals)
+        else:
+            self.environment = environment
         self.error_tasks = {}
 
     def validate(self, expression):
@@ -175,7 +120,7 @@ class PythonScriptEngine(object):
         same name as a pre-defined script, rending the script un-callable.
         This results in a nearly indecipherable error.  Better to fail
         fast with a sensible error message."""
-        func_overwrites = set(self.globals).intersection(task.data)
+        func_overwrites = set(self.environment.globals).intersection(task.data)
         func_overwrites.update(set(external_methods).intersection(task.data))
         if len(func_overwrites) > 0:
             msg = f"You have task data that overwrites a predefined " \
@@ -183,45 +128,8 @@ class PythonScriptEngine(object):
                   f"field name(s) to something else: {func_overwrites}"
             raise WorkflowTaskException(msg, task=task)
 
-    def convert_to_box(self, data):
-        if isinstance(data, dict):
-            for key, value in data.items():
-                if not isinstance(value, Box):
-                    data[key] = self.convert_to_box(value)
-            return Box(data)
-        if isinstance(data, list):
-            for idx, value in enumerate(data):
-                data[idx] = self.convert_to_box(value)
-            return data
-        return data
-
     def _evaluate(self, expression, context, external_methods=None):
-
-        my_globals = copy.copy(self.globals)  # else we pollute all later evals.
-        self.convert_to_box(context)
-        my_globals.update(external_methods or {})
-        my_globals.update(context)
-        return eval(expression, my_globals)
+        return self.environment.evaluate(expression, context, external_methods)
 
     def _execute(self, script, context, external_methods=None):
-
-        my_globals = copy.copy(self.globals)
-        self.convert_to_box(context)
-        my_globals.update(external_methods or {})
-        context.update(my_globals)
-        try:
-            exec(script, context)
-        finally:
-            self.remove_globals_and_functions_from_context(context,
-                                                           external_methods)
-
-    def remove_globals_and_functions_from_context(self, context,
-                                                  external_methods=None):
-        """When executing a script, don't leave the globals, functions
-        and external methods in the context that we have modified."""
-        for k in list(context):
-            if k == "__builtins__" or \
-                    hasattr(context[k], '__call__') or \
-                    k in self.globals or \
-                    external_methods and k in external_methods:
-                context.pop(k)
+        self.environment.execute(script, context, external_methods)
