+++ tests/SpiffWorkflow/dmn/feel_engine/FeelDictDecisionTest.py
@@ -1,6 +1,6 @@
 import unittest
 
-from SpiffWorkflow.bpmn.PythonScriptEngine import PythonScriptEngine
+from SpiffWorkflow.bpmn.PythonScriptEngineEnvironment import Box
 
 from .FeelDecisionRunner import FeelDecisionRunner
 
@@ -19,7 +19,7 @@ class FeelDictDecisionTestClass(unittest.TestCase):
                 "PEANUTS": {"delicious": True},
                 "SPAM": {"delicious": False}
                 }}
-        PythonScriptEngine.convert_to_box(PythonScriptEngine(), data)
+        Box.convert_to_box(data)
         res = self.runner.decide(data)
         self.assertEqual(res.description, 'They are allergic to peanuts')
